requirements.txt
antlr4-python3-runtime==4.8
click==8.0.1
joblib==1.0.1
nltk==3.6.2
omegaconf==2.1.0
PyStemmer==2.0.1
PyYAML==5.4.1
regex==2021.8.3
termcolor==1.1.0
tqdm==4.62.1
